requirements.txt
faiss-cpu
pdfminer.six
PyGithub
requests
numpy
pyfiglet
termcolor
colorama
python-dotenv